+++ lib/src/data/personal/agama.dart
@@ -1,0 +1,1 @@
+const listAgama = ["Islam", "Kristen", "Katolik", "Hindu", "Budha", "Konghucu"];