+++ lib/src/data/warna/warna.dart
@@ -1,0 +1,81 @@
+const listWarna = [
+  /// A
+  "Abu-abu",
+  "Abu-abu Beton",
+  "Abu-abu Terang",
+  "Abu-abu Tua",
+
+  /// B
+  "Beige",
+  "Biru",
+  "Biru Matang",
+  "Biru Muda",
+  "Biru Langit",
+  "Biru Laut",
+  "Biru Tua",
+
+  /// Coklat
+  "Coklat",
+  "Coklat Muda",
+  "Coklat Tua",
+
+  /// Emas
+  "Emas",
+
+  /// H
+  "Hijau",
+  "Hijau Botol",
+  "Hijau Daun",
+  "Hijau Lumut",
+  "Hijau Mint",
+  "Hijau Stabilo",
+  "Hijau Tua",
+  "Hitam",
+
+  /// I
+  "Indigo",
+
+  /// J
+  "Jeruk",
+  "Jingga",
+
+  /// K
+  "Krem",
+  "Kuning",
+  "Kuning Lemon",
+  "Kuning Pasir",
+  "Kuning Telur",
+
+  /// L
+  "Lavender",
+
+  /// M
+  "Marun",
+  "Merah",
+  "Merah Anggur",
+  "Merah Bata",
+  "Merah Jambu",
+  "Merah Merapi",
+  "Merah Muda",
+  "Merah Tua",
+
+  // N
+  "Nila",
+
+  /// P
+  "Perak",
+  "Putih",
+  "Putih Tulang",
+
+  /// S
+  "Silver",
+
+  /// U
+  "Ungu",
+  "Ungu Gelap",
+  "Ungu Muda",
+  "Ungu Tua",
+
+  /// T
+  "Tosca",
+];